planner: move physical PhysicalSequence related logic into physical op dir.
